+++ tendrl/node_agent/flows/__init__.py
@@ -1,0 +1,10 @@
+import logging
+
+from tendrl.commons import flows
+
+LOG = logging.getLogger(__name__)
+
+
+class NodeAgentBaseFlow(flows.BaseFlow):
+    def __init__(self, *args, **kwargs):
+        super(NodeAgentBaseFlow, self).__init__(*args, **kwargs)
